feat：调整项目接口逻辑

+++ server-nest/src/modules/roles/roles.service.ts
@@ -26,12 +26,35 @@ export class RolesService {
     return createdRole.save();
   }
 
-  async findAll(query?: any): Promise<Role[]> {
-    const filter = {};
-    if (query?.status !== undefined) {
-      filter['status'] = query.status;
+  async findAll(query?: any): Promise<any> {
+    const params = query?.params || {};
+    const current = Number(query?.pagination?.current || 1) || 1;
+    const pageSize = Number(query?.pagination?.pageSize || 15) || 15;
+    const sortColumn = query?.sort?.columnKey || 'createdAt';
+    const sortOrder = query?.sort?.order === 'ascend' ? 1 : -1;
+
+    const fuzzyParams: Record<string, any> = {};
+    for (const key in params) {
+      if (params.hasOwnProperty(key) && params[key]) {
+        fuzzyParams[key] = { $regex: new RegExp(params[key], 'i') };
+      }
     }
-    return this.roleModel.find(filter).exec();
+
+    const aggregationPipeline: any[] = [
+      { $match: fuzzyParams },
+      { $sort: { [sortColumn]: sortOrder } },
+      { $skip: (current - 1) * pageSize },
+      { $limit: pageSize },
+    ];
+
+    const [result, total] = await Promise.all([
+      this.roleModel.aggregate(aggregationPipeline).exec(),
+      this.roleModel.countDocuments(params).exec(),
+    ]);
+
+    return result.length > 0
+      ? { result, current, pageSize, total }
+      : { result: [], total };
   }
 
   async findOne(id: string): Promise<Role> {
@@ -51,6 +74,14 @@ export class RolesService {
       throw new NotFoundException('无效的角色ID');
     }
 
+    const roleInfo = await this.roleModel.findById(id).exec();
+    if (!roleInfo) {
+      throw new NotFoundException('角色不存在');
+    }
+    if (roleInfo.roleAuth === 'SUPER-ADMIN') {
+      throw new ConflictException('不能修改超级管理员');
+    }
+
     const updatedRole = await this.roleModel
       .findByIdAndUpdate(id, updateRoleDto, { new: true })
       .exec();
@@ -67,9 +98,17 @@ export class RolesService {
       throw new NotFoundException('无效的角色ID');
     }
 
-    const result = await this.roleModel.findByIdAndDelete(id).exec();
-    if (!result) {
+    const role = await this.roleModel.findById(id).exec();
+    if (!role) {
       throw new NotFoundException('角色不存在');
     }
+    if (role.roleAuth === 'SUPER-ADMIN') {
+      throw new ConflictException('不能删除超级管理员角色');
+    }
+    if (role.roleAuth === 'VISITOR-ADMIN') {
+      throw new ConflictException('不能删除访客管理员角色');
+    }
+
+    await this.roleModel.findByIdAndDelete(id).exec();
   }
 }
